+++ server/Recruit-Domain/src/main/java/com/econovation/recruitdomain/domains/dto/RecruitmentSetUpDto.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class RecruitmentSetUpDto {
 
-    private Long year;
+    private Integer year;
     private Long startAt;
     private Long endAt;
